scudo: Create a public include directory. NFCI.
Summary:
For MTE error reporting we will need to expose interfaces for crash handlers
to use to interpret scudo headers and metadata. The intent is that these
interfaces will live in scudo/interface.h.

Move the existing interface.h into an include/scudo directory and make it
independent of the internal headers, so that we will be able to add the
interfaces there.

Reviewers: cryptoad

Subscribers: mgorny, #sanitizers, hctim, eugenis, llvm-commits

Tags: #sanitizers

Differential Revision: https://reviews.llvm.org/D76648