[AArch64][MachO] Support @AUTH relocation.
This adds parser, MCExpr, and emitter support for the arm64e-specific
authenticated pointer auth relocation.

In assembly, this is expressed using:
  .quad <symbol>@AUTH(<key>, <discriminator> [, addr])
For example:
  .quad _g3@AUTH(ib, 1234, addr)

The optional 'addr' specifier represents whether the generated pointer
authentication code will also include address diversity (by blending the
address of the storage location of the relocated pointer with the
user-specified constant discriminator).

The @AUTH expression lowers to a new kind of MachO relocation:
  ARM64_RELOC_AUTHENTICATED_POINTER
It points to the referenced symbol, and the other data, describing
the signing scheme and original addend (shrunk to 32 bits instead of 64),
is encoded into the addend (in the relocated location), as such:

  |63|62|61-51|50-49|  48  |47     -     32|31  -  0|
  | 1| 0|  0  | key | addr | discriminator | addend |