WIP: CodeGen: Introduce a new "interop" encoding for CFI function types.
The goal of this encoding is to make it easier for foreign languages to
interoperate with Clang CFI by allowing both languages to produce
the same function types. This is achieved by removing as many
language-specific details from the encoded function types as possible
while still providing some degree of type checking.

TODO:
- Introduce a separate flag for this instead of changing the behavior of
  -fsanitize-cfi-icall-generalize-pointers
- Tests, docs