scudo: Split up SCUDO_ANDROID and SCUDO_BIONIC.
SCUDO_ANDROID indicates that we are running on an Android kernel, while
SCUDO_BIONIC indicates that libc is Bionic. Usually they are the same,
but when Bionic is being compiled for the host SCUDO_ANDROID is false
and SCUDO_BIONIC is true. When targeting Bionic we generally want to
make the same decisions whether we are running on an Android kernel
or not, but when SCUDO_ANDROID is false we need to be compatible with
non-Android kernels. In particular, we cannot assume that the page size
is 4096 and that PR_SET_VMA is available.

Therefore, create a SCUDO_BIONIC macro and use it everywhere except
for the page size assumption and PR_SET_VMA prctl() call, where we
use SCUDO_ANDROID.